services/ingestA/pyproject.toml
[project]
name = "ingesta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chevron>=0.14.0",
    "fastapi>=0.135.1",
    "google-cloud-bigquery>=3.40.1",
    "netaddr>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "tzlocal>=5.3.1",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
